+++ CHANGES.rst
@@ -202,6 +202,7 @@ Accessibility
 ^^^^^^^^^^^^^
 
 - Improve registration form single choice input accessibility (:pr:`6310`, thanks :user:`foxbunny`)
+- Make error messages accessible in the registration form (:pr:`6324`, thanks :user:`foxbunny`)
 
 Internal Changes
 ^^^^^^^^^^^^^^^^